refactor(bot/application): change interfaces to application-interfaces
