+++ booking-telegram-bot/contexts/telegram-bot/application-module/src/index.ts
@@ -1,3 +1,4 @@
 export * from './module/index.js'
 export * from './ports/index.js'
 export * from './use-cases/index.js'
+export * from './interfaces/index.js'
