+++ booking-telegram-bot/contexts/telegram-bot/application-module/src/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './telegram-client.interfaces.js'
