+++ booking-telegram-bot/contexts/telegram-bot/application-module/src/interfaces/telegram-client.interfaces.ts
