+++ booking-telegram-bot/contexts/telegram-bot/application-module/src/ports/telegram-client.port.ts
@@ -1,6 +1,5 @@
-import type { CreateConversationReturnType }    from '@booking-telegram-bot/tgsnake-adapter'
-
-import type { TelegramBotFormattedContextType } from './telegram-client.interfaces.js'
+import type { CreateConversationReturnType }    from '@telegram-bot/application-module'
+import type { TelegramBotFormattedContextType } from '@telegram-bot/application-module'
 
 export abstract class TelegramClientPort {
   abstract sendMessage(ctx: TelegramBotFormattedContextType, text: string): Promise<void>
