+++ booking-telegram-bot/contexts/telegram-bot/application-module/src/services/appointment-conversation/index.ts
@@ -1,2 +1,2 @@
 export * from './appointment-conversation.service.js'
-export * from './conversation-parts/index.js'
+export * from './conversation-q-a-pairs/index.js'
